ld: Add DT_FLAGS_2 and DF_2_GNU_IFUNC
As shown in

https://sourceware.org/bugzilla/show_bug.cgi?id=20019

it is possible to create a shared object which references an IFUNC
function defined in another shared object.  With non-lazy binding at
run-time, usage of such a shared object can lead to segfault.  The
issue is the shared object, which provides the IFUNC function, isn't
in DT_NEEDED and dynamic linker tries to resolve the reference to the
IFUNC function defined in the unrelocated shared object with non-lazy
binding.  Currently, the glibc dynamic linker issues a warning:

[hjl@gnu-cfl-1 pr20019]$ ./main-dynamic
./main-dynamic: Relink `./libbar.so' with `/export/build/gnu/glibc/build-x86_64-linux/libc.so.6' for IFUNC symbol `memmove'
Segmentation fault
[hjl@gnu-cfl-1 pr20019]$

But it doesn't prevent segfault.

This patch extends gABI with DT_FLAGS_2 and DF_2_GNU_IFUNC to indicate
that a shared object or executable has IFUNC symbols so that dynamic
linker can relocate it early.

bfd/

	* elflink.c (bfd_elf_size_dynamic_sections): Set DF_2_GNU_IFUNC
	if there are GNU IFUNC symbols in shared objects or executables
	with interpreter.

binutils/

	* readelf.c (get_dynamic_type): Handle DT_FLAGS_2.
	(process_dynamic_section): Handle DT_FLAGS_2 and DF_2_GNU_IFUNC.

include/

	* bfdlink.h (bfd_link_info): Add flags_2.
	* elf/common.h (DT_FLAGS_2): New.
	(DF_2_GNU_IFUNC): Likewise.

ld/

	* testsuite/ld-ifunc/ifunc-2-i386-now.d: Updated.
	* testsuite/ld-ifunc/ifunc-2-local-i386-now.d: Likewise.
	* testsuite/ld-ifunc/ifunc-2-local-x86-64-now.d: Likewise.
	* testsuite/ld-ifunc/ifunc-2-x86-64-now.d: Likewise.
	* testsuite/ld-ifunc/pr17154-x86-64-now.d: Likewise.
	* testsuite/ld-ifunc/pr17154-x86-64.d: Likewise.
	* testsuite/ld-x86-64/bnd-ifunc-1-now.d: Likewise.
	* testsuite/ld-x86-64/bnd-ifunc-2-now.d: Likewise.
	* testsuite/ld-x86-64/bnd-ifunc-2.d: Likewise.
	* ld-x86-64/pr20253-1d.d: Likewise.
	* ld-x86-64/pr20253-1f.d: Likewise.
	* ld-x86-64/pr20253-1j.d: Likewise.
	* ld-x86-64/pr20253-1l.d: Likewise.
	* testsuite/ld-ifunc/ifunc-26.d: Pass -d to readelf.  Check
	for DF_2_GNU_IFUNC.